feat: redis 를 사용한 시청자 수 계산

+++ chat-backend/src/main/java/com/example/chatbackend/controller/ViewerCountController.java
@@ -14,7 +14,7 @@ public class ViewerCountController {
     }
 
     @GetMapping("/viewer-count")
-    public int getViewerCount() {
+    public Long getViewerCount() {
         return chatWebSocketHandler.getViewerCount();
     }
 }
